Property: Make moved Signal private
The moved Signal is problematic, as we can no longer guarantee noexcept
move behavior of Properties, if anyone is allowed to connect a slot to
the moved Signal. (See #24)

By making the moved Signal private within KDBindings, we can have
control over what code is executed by it. This allows us to keep the
noexcept guarantee on the Property move constructor/assignment.

This also warrants a version bump to v1.0.1