Introduce ScopedConnectionHandle
Currently we have to disconnect the slot ourselves when its not needed.
This is a bit error prone as we might forget to disconnect a slot in
some cases leading to use after frees and invalid reads.

This change proposes ScopedConnectionHandle which is a thin wrapper
over ConnectionHandle. It automatically disconnects on scope exit.