ci: Use windows-2019 instead of -latest
We've been dealing with an issue in CI, where the test-signal test case
fails when building in Release mode on Windows.
It fails in the ConnectionEvaluator, when it tries to lock the mutex,
which is something that should simply not happen, as the shared_ptr in
the testcase keeps the ConnectionEvaluator alive at the point where the
crash occurs.

In addition, this problem is not reproducable anywhere outside Githubs
CI. If I copy the executable from Github Actions to my local machine,
it simply works.
This leads me to believe this is only a problem with this very specific
Windows environment.
By downgrading our Windows version in CI, we can avoid the problem and
at least continue development.