feat: ConnectionHandle::release to ignore warning
As described in #43, you now get a warning if you don't use the
ConnectionHandle that is returned by connecting to a signal.
But as @lemirep noted, this can be annoying without a proper way to
silence the warning.
This commit adds a `release` function that can simply be used to ignore
the nodiscard warning intentionally.

Fix #43