Add example CMakeUserPresets.json file
Useful on Windows if you have the LLVM distro installed for tools
such as clang-format. When this is in the path the standard "dev"
preset finds clang in preference to cl when configuring. The example
"msvc_x64-dev" helps to find cl.

Just copy and rename to CMakeUserPresets.json to use and customize.

Added to .gitignore to prevent accidental commits of local modifications.